+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added JSX/TSX support to RxFM.
 
+### Changed
+- Updated TypeScript version to ^4.5.4, removed deprecated `applet` html component creator.
+
 ### Deprecated
 - Deprecated `DestroySubject`.
 - Deprecated `ConditionalOptions` overload in `conditional` operator.